feat | Add function logout
