+++ app/components/ui/index.ts
@@ -1,4 +1,3 @@
-export * from './sidebarItem/SidebarItem';
 export * from './card/CardWithIcon';
 export * from './horizontalBar/HorizontalBar';
 export * from './table/TableWithFilter';
