+++ app/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./navbar/Navbar";
+export * from "./sidebar/Sidebar";
